+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.7
 
 - Support `stroke` `:dashes`
+- Support symbol values for draw_line_cap (`:round`, `:square`, `:flat`) and draw_line_join (`:miter`, `:round`, `:bevel`) and draw_fill_mode (`:winding` and `:alternate`)
 
 ## 0.1.6